backend/requirements.txt
neo4j>=5.0.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
